Split datatype from wbi_core to wbi_datatype
