+++ test/test_wbi_core.py
@@ -1,6 +1,6 @@
 import unittest
 
-from wikibaseintegrator import wbi_core, wbi_functions
+from wikibaseintegrator import wbi_core, wbi_functions, wbi_datatype
 
 
 class TestWbiCore(unittest.TestCase):
@@ -9,8 +9,8 @@ class TestWbiCore(unittest.TestCase):
     def test_item_engine(self):
         wbi_core.ItemEngine(debug=True)
         wbi_core.ItemEngine(data=None, debug=True)
-        wbi_core.ItemEngine(data=wbi_core.String(value='test', prop_nr='P1'), debug=True)
-        wbi_core.ItemEngine(data=[wbi_core.String(value='test', prop_nr='P1')], debug=True)
+        wbi_core.ItemEngine(data=wbi_datatype.String(value='test', prop_nr='P1'), debug=True)
+        wbi_core.ItemEngine(data=[wbi_datatype.String(value='test', prop_nr='P1')], debug=True)
         with self.assertRaises(TypeError):
             wbi_core.ItemEngine(data='test', debug=True)
         with self.assertRaises(ValueError):
@@ -34,10 +34,10 @@ class TestWbiCore(unittest.TestCase):
         assert item.get_label("es") == "Tierra"
 
     def test_basedatatype_if_exists(self):
-        instance_of_append = wbi_core.ItemID(prop_nr='P31', value='Q1234', if_exists='APPEND')
-        instance_of_forceappend = wbi_core.ItemID(prop_nr='P31', value='Q1234', if_exists='FORCE_APPEND')
-        instance_of_replace = wbi_core.ItemID(prop_nr='P31', value='Q1234', if_exists='REPLACE')
-        instance_of_keep = wbi_core.ItemID(prop_nr='P31', value='Q1234', if_exists='KEEP')
+        instance_of_append = wbi_datatype.ItemID(prop_nr='P31', value='Q1234', if_exists='APPEND')
+        instance_of_forceappend = wbi_datatype.ItemID(prop_nr='P31', value='Q1234', if_exists='FORCE_APPEND')
+        instance_of_replace = wbi_datatype.ItemID(prop_nr='P31', value='Q1234', if_exists='REPLACE')
+        instance_of_keep = wbi_datatype.ItemID(prop_nr='P31', value='Q1234', if_exists='KEEP')
 
         item = wbi_core.ItemEngine(item_id="Q2", data=[instance_of_append, instance_of_append])
         claims = [x['mainsnak']['datavalue']['value']['id'] for x in item.get_json_representation()['claims']['P31']]
@@ -119,38 +119,38 @@ class TestWbiCore(unittest.TestCase):
 
     def test_new_item_creation(self):
         data = [
-            wbi_core.String(value='test1', prop_nr='P1'),
-            wbi_core.String(value='test2', prop_nr='1'),
-            wbi_core.String(value='test3', prop_nr=1),
-            wbi_core.Math("xxx", prop_nr="P2"),
-            wbi_core.ExternalID("xxx", prop_nr="P3"),
-            wbi_core.ItemID("Q123", prop_nr="P4"),
-            wbi_core.ItemID("123", prop_nr="P4"),
-            wbi_core.ItemID(123, prop_nr="P4"),
-            wbi_core.Time(time='-0458-00-00T00:00:00Z', before=1, after=2, precision=3, timezone=4, prop_nr="P5"),
-            wbi_core.Time(time='458-00-00T00:00:00Z', before=1, after=2, precision=3, timezone=4, prop_nr="P5"),
-            wbi_core.Time(time='+2021-01-01T15:15:15Z', before=1, after=2, precision=3, timezone=4, prop_nr="P5"),
-            wbi_core.Url("http://www.wikidata.org", prop_nr="P6"),
-            wbi_core.Url("https://www.wikidata.org", prop_nr="P6"),
-            wbi_core.Url("ftp://example.com", prop_nr="P6"),
-            wbi_core.Url("ssh://user@server/project.git", prop_nr="P6"),
-            wbi_core.Url("svn+ssh://user@server:8888/path", prop_nr="P6"),
-            wbi_core.MonolingualText(text="xxx", language="fr", prop_nr="P7"),
-            wbi_core.Quantity(quantity=-5.04, prop_nr="P8"),
-            wbi_core.Quantity(quantity=5.06, upper_bound=9.99, lower_bound=-2.22, unit="Q11573", prop_nr="P8"),
-            wbi_core.CommonsMedia("xxx", prop_nr="P9"),
-            wbi_core.GlobeCoordinate(latitude=1.2345, longitude=-1.2345, precision=12, prop_nr="P10"),
-            wbi_core.GeoShape("Data:xxx.map", prop_nr="P11"),
-            wbi_core.Property("P123", prop_nr="P12"),
-            wbi_core.Property("123", prop_nr="P12"),
-            wbi_core.Property(123, prop_nr="P12"),
-            wbi_core.TabularData("Data:Taipei+Population.tab", prop_nr="P13"),
-            wbi_core.MusicalNotation("\relative c' { c d e f | g2 g | a4 a a a | g1 |}", prop_nr="P14"),
-            wbi_core.Lexeme("L123", prop_nr="P15"),
-            wbi_core.Lexeme("123", prop_nr="P15"),
-            wbi_core.Lexeme(123, prop_nr="P15"),
-            wbi_core.Form("L123-F123", prop_nr="P16"),
-            wbi_core.Sense("L123-S123", prop_nr="P17")
+            wbi_datatype.String(value='test1', prop_nr='P1'),
+            wbi_datatype.String(value='test2', prop_nr='1'),
+            wbi_datatype.String(value='test3', prop_nr=1),
+            wbi_datatype.Math("xxx", prop_nr="P2"),
+            wbi_datatype.ExternalID("xxx", prop_nr="P3"),
+            wbi_datatype.ItemID("Q123", prop_nr="P4"),
+            wbi_datatype.ItemID("123", prop_nr="P4"),
+            wbi_datatype.ItemID(123, prop_nr="P4"),
+            wbi_datatype.Time(time='-0458-00-00T00:00:00Z', before=1, after=2, precision=3, timezone=4, prop_nr="P5"),
+            wbi_datatype.Time(time='458-00-00T00:00:00Z', before=1, after=2, precision=3, timezone=4, prop_nr="P5"),
+            wbi_datatype.Time(time='+2021-01-01T15:15:15Z', before=1, after=2, precision=3, timezone=4, prop_nr="P5"),
+            wbi_datatype.Url("http://www.wikidata.org", prop_nr="P6"),
+            wbi_datatype.Url("https://www.wikidata.org", prop_nr="P6"),
+            wbi_datatype.Url("ftp://example.com", prop_nr="P6"),
+            wbi_datatype.Url("ssh://user@server/project.git", prop_nr="P6"),
+            wbi_datatype.Url("svn+ssh://user@server:8888/path", prop_nr="P6"),
+            wbi_datatype.MonolingualText(text="xxx", language="fr", prop_nr="P7"),
+            wbi_datatype.Quantity(quantity=-5.04, prop_nr="P8"),
+            wbi_datatype.Quantity(quantity=5.06, upper_bound=9.99, lower_bound=-2.22, unit="Q11573", prop_nr="P8"),
+            wbi_datatype.CommonsMedia("xxx", prop_nr="P9"),
+            wbi_datatype.GlobeCoordinate(latitude=1.2345, longitude=-1.2345, precision=12, prop_nr="P10"),
+            wbi_datatype.GeoShape("Data:xxx.map", prop_nr="P11"),
+            wbi_datatype.Property("P123", prop_nr="P12"),
+            wbi_datatype.Property("123", prop_nr="P12"),
+            wbi_datatype.Property(123, prop_nr="P12"),
+            wbi_datatype.TabularData("Data:Taipei+Population.tab", prop_nr="P13"),
+            wbi_datatype.MusicalNotation("\relative c' { c d e f | g2 g | a4 a a a | g1 |}", prop_nr="P14"),
+            wbi_datatype.Lexeme("L123", prop_nr="P15"),
+            wbi_datatype.Lexeme("123", prop_nr="P15"),
+            wbi_datatype.Lexeme(123, prop_nr="P15"),
+            wbi_datatype.Form("L123-F123", prop_nr="P16"),
+            wbi_datatype.Sense("L123-S123", prop_nr="P17")
         ]
         core_props = set(["P{}".format(x) for x in range(20)])
